+++ frontend/src/components/molecules/cards/Frosted/Frosted.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Card } from 'antd';
+
+import './Frosted.css';
+
+function Frosted() {
+  return (
+    <Card id="frostedcard" />
+  );
+}
+
+export default Frosted;
