+++ frontend/src/components/organisms/TopLevelSearchContainer/TopLevelSearchContainer.css
@@ -1,0 +1,10 @@
+#frostedcard {
+    border: none;
+    border-radius: 4px;
+    background: rgba(255, 245, 255, 0.2);
+    backdrop-filter: blur(3px);
+    margin: auto;
+    margin-bottom: 70px;
+    width: 50%;
+    min-width: 650px;
+}
