+++ frontend/src/components/organisms/TopLevelSearchContainer/TopLevelSearchContainer.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import FrostedCard from '../../molecules/cards/Frosted/Frosted';
+import TopLevelSearchForm from '../TopLevelSearchForm/TopLevelSearchForm';
+
+function TopLevelSearchContainer() {
+  return (
+    <FrostedCard>
+      <TopLevelSearchForm />
+    </FrostedCard>
+  );
+}
+
+export default TopLevelSearchContainer;
