--- frontend/src/components/organisms/TopLevelSearchFormFields/TopLevelSearchFormFields.css
@@ -1,69 +1,0 @@
-@import '~antd/dist/antd.dark.css';
-
-.ant-input {
-    color: white !important;
-    background-color: rgba(255, 255, 255, 0.15) !important;
-    border: 1px solid #939393 !important;
-}
-
-.ant-select:not(.ant-select-customize-input) .ant-select-selector {
-    color: white !important;
-    background-color: rgba(255, 255, 255, 0.15) !important;
-    border: 1px solid #939393 !important;
-}
-
-.ant-select-arrow {
-    color: whitesmoke !important;
-}
-
-.ant-checkbox-group {
-    display: inline !important;
-}
-
-.ant-checkbox-wrapper {
-    color: whitesmoke !important;
-}
-
-/* .ant-select-single.ant-select-open .ant-select-selection-item {
-    color: white !important;
-    background-color: rgba(255, 255, 255, 0.15) !important;
-    border: 1px solid #939393 !important;
-} */
-
-.ant-form-item-label > label {
-    color: whitesmoke !important;
-}
-
-.ant-form-item-label > label .ant-form-item-tooltip {
-    color: whitesmoke !important;
-}
-
-.ant-form-item-label > label .ant-form-item-optional {
-    color: whitesmoke !important;
-}
-
-.ant-collapse {
-    border: 1px solid #939393 !important;
-}
-
-.ant-collapse-content {
-    color: whitesmoke !important;
-    background-color: rgba(27, 39, 53, 0.95) !important;
-}
-
-.ant-collapse > .ant-collapse-item > .ant-collapse-header {
-    color: white !important;
-    background-color: rgba(27, 39, 53, 0.85) !important;
-}
-
-/* .ant-radio-button-wrapper {
-    border: 1px solid #939393;
-}
-
-.ant-radio-button-wrapper:first-child {
-    border-left: 1px solid #939393;
-} */
-
-/* .ant-radio-button-wrapper:last-child {
-    border-left: 1px solid #939393 !important;
-} */
